server/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
pytest
httpx
python-dotenv
pywebpush
yfinance>=0.2.0
google-genai>=1.0.0
scikit-learn
supabase>=2.0.0
